requirements.txt
python-docx==0.8.11
requests==2.27.1
PyQt5==5.9.7